Properly handle directory components in filenames
When searching for a file using a search path, there are three
possibilities. Say the file to find is /rootfiles/exp/data.root.
Then:
path             filename
                 /rootfiles/exp/data.root
/rootfiles       exp/data.root
/rootfiles/exp   data.root

In all three cases, we let OpenInPath return /rootfiles/exp as the
"found path" (fp). The full path to the file is then always
fp + "/" + BasenameStr(filename).