Support specifying ROOT file and path on via command line options
As in the original panguin, we ignore the root file name from either command
line or configuration file if a run number is specified at the same time. In
this case, a warning is printed. We also print a warning if the ROOT file name
appears to have a different run number than the one given on the command line,
to alert users to potential mistakes.