Honor -nostat option for macro and tree draw commands and saved images
This may not always work with a macro since it can override the stat
box settings.

Note that ROOT does not automatically generate a statistics box for 2D
(or 3D) plots generated with TTree::Draw. See
https://root.cern.ch/root/roottalk/roottalk01/1084.html.
In other words, a draw command like
 treevar1:treevar2
will not produce a stat box, even without "-nostat".
To get the statistics box for a 2D plot, do
 treevar1:treevar2>>myh2f
or similar.